+++ CHANGELOG.md
@@ -7,12 +7,13 @@ This file contains the list of changes made to pyjoulescope_ui.
 
 ## 1.1.0
 
-2024 Mar 9 [in progress]
+2024 Mar 18 [in progress]
 
 * Added plugin framework with live code reload  #14
 * Restructured code to support human language translation.
 * Added AI translations: ar, de, el, es, fr, ja, ko, zh.
 * Added Intel graphics dialog to display sequence  #245
+* Added Trigger widget.  [work in progress]
 
 
 ## 1.0.59